Convert value to grc units in getstakinginfo

+++ src/miner.h
@@ -14,7 +14,7 @@ struct CMinerStatus
     CCriticalSection lock;
     std::string ReasonNotStaking;
     std::string Message;
-    double WeightSum,WeightMin,WeightMax;
+    uint64_t WeightSum,WeightMin,WeightMax;
     double ValueSum;
     double CoinAgeSum;
     int Version;
